+++ NEWS.md
@@ -1,5 +1,11 @@
 # simmer devel
 
+## New features
+
+- Add support for subsetting by activity tag (#305).
+
+## Minor changes and fixes
+
 - Fix `set_source()` to avoid leaking arrivals from the old source (#322).
 - Fix sources to properly reset distributions and trajectories (#324).
 - Fix resources to properly reset initial parameters (#325).